+++ wp-content/uploads/civicrm/ext/firewall/Civi/Firewall/Listener/DeclinedCard.php
@@ -10,9 +10,10 @@
  */
 namespace Civi\Firewall\Listener;
 
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Civi\Api4\FirewallIpaddress;
+use Civi\Core\Service\AutoSubscriber;
 
-class DeclinedCard implements EventSubscriberInterface {
+class DeclinedCard extends AutoSubscriber {
 
   /**
    * @return array
@@ -28,8 +29,7 @@ class DeclinedCard implements EventSubscriberInterface {
 
   public function onTrigger(\Civi\Firewall\Event\DeclinedCardEvent $event) {
     // Add to firewall ip address log table with timestamp + event type
-    \Civi\Api4\FirewallIpaddress::create()
-      ->setCheckPermissions(FALSE)
+    FirewallIpaddress::create(FALSE)
       ->addValue('ip_address', $event->ipAddress)
       ->addValue('source', $event->source)
       ->addValue('event_type', $event->eventType)